+++ web/apps/photos/src/components/PhotoViewer/ImageEditorOverlay/index.tsx
@@ -7,11 +7,11 @@ import {
 import type { MiniDialogAttributes } from "@/base/components/MiniDialog";
 import { nameAndExtension } from "@/base/file";
 import log from "@/base/log";
+import { downloadAndRevokeObjectURL } from "@/base/utils/web";
 import { EnteFile } from "@/media/file";
 import { photosDialogZIndex } from "@/new/photos/components/z-index";
 import downloadManager from "@/new/photos/services/download";
 import { AppContext } from "@/new/photos/types/context";
-import { downloadAndRevokeObjectURL } from "@/new/photos/utils/web";
 import { ensure } from "@/utils/ensure";
 import {
     CenteredFlex,
