+++ web/apps/photos/src/components/Sidebar/index.tsx
@@ -7,11 +7,12 @@ import { useModalVisibility } from "@/base/components/utils/modal";
 import log from "@/base/log";
 import { savedLogs } from "@/base/log-web";
 import { customAPIHost } from "@/base/origins";
+import { downloadString } from "@/base/utils/web";
 import { RecoveryKey } from "@/new/photos/components/RecoveryKey";
 import { downloadAppDialogAttributes } from "@/new/photos/components/utils/download";
 import type { CollectionSummaries } from "@/new/photos/services/collection/ui";
 import { AppContext, useAppContext } from "@/new/photos/types/context";
-import { downloadString, initiateEmail, openURL } from "@/new/photos/utils/web";
+import { initiateEmail, openURL } from "@/new/photos/utils/web";
 import { SpaceBetweenFlex } from "@ente/shared/components/Container";
 import { EnteMenuItem } from "@ente/shared/components/Menu/EnteMenuItem";
 import ThemeSwitcher from "@ente/shared/components/ThemeSwitcher";
